+++ app/controllers/site_controller.rb
@@ -1,4 +1,18 @@
 class SiteController < ApplicationController
-  def index
+  before_action :authenticate_user!
+
+  def index; end
+
+  def home; end
+
+  def profile; end
+
+  def dashboard
+    @features = current_user.features.all
+    platforms = []
+    @features.each do |f|
+      platforms << f.service.platform
+    end
+    @platforms = platforms.uniq
   end
 end