+++ app/models/platform.rb
@@ -1,0 +1,5 @@
+class Platform < ApplicationRecord
+  has_many :services, dependent: :destroy
+
+  validates :name, presence: true
+end