+++ app/models/service.rb
@@ -1,0 +1,6 @@
+class Service < ApplicationRecord
+  belongs_to :platform
+  has_many :features, dependent: :destroy
+
+  validates :name, presence: true
+end