+++ db/migrate/20230502013013_remove_column_from_feature.rb
@@ -1,0 +1,7 @@
+class RemoveColumnFromFeature < ActiveRecord::Migration[7.0]
+  def change
+    remove_column :features, :url
+    remove_column :features, :emoji
+    remove_column :features, :name
+  end
+end