+++ CHANGELOG.md
@@ -3,7 +3,15 @@ All notable changes to this project will be documented in this file.
  
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## [0.2.0] - 2023-06-06
  
+### Added
+- Added functionality to change avatar
+- Made the Dashboard functionality dynamic
+- Created a Home page
+- Created a Profile page 
+
 ## [0.1.0] - 2023-04-28
  
 ### Added